refactor: useUpdatePlaylist
